doc_pipeline/requirements.txt
# OCR Dependencies
easyocr>=1.7.0
Pillow>=10.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0

# AI/LLM Dependencies
google-generativeai>=0.8.0
python-dotenv>=1.0.0

# Core Dependencies
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0

# Utilities
pathlib2>=2.3.7
typing-extensions>=4.0.0

# PDF Extraction
PyMuPDF>=1.23.0
pdfplumber>=0.10.2

# CLI
typer>=0.9.0
rich>=13.0.0 